DRbCached client implements consistent hashing of keys across clusters.
The client is initialized with a Server or list of Servers. It computes
the hash of the Server object's string and maps the hash to ranges on
between 0 and 0xFFFF_FFFF.

When the client wants to read or write a value to the cluster, it hashes
the key and writes to the first node whose hash >= hash(key).